+++ assets/jsx/components/DateOffsetPreview.jsx
@@ -5,6 +5,7 @@ import {
     Fragment
 } from '@wordpress/element';
 import { addQueryArgs } from '@wordpress/url';
+import { __ } from '@wordpress/i18n';
 
 const { apiFetch } = wp;
 
@@ -52,7 +53,7 @@ export const DateOffsetPreview = ({
                 setValidationErrorCallback(result.message);
 
                 if (result.isValid) {
-                    setOffsetPreview(result.preview);
+                    setOffsetPreview(result.calculatedTime);
                     setCurrentTime(result.currentTime);
                 } else {
                     setOffsetPreview('');